+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## DEPENDENCIES
 
+- **grunt-contrib-compass** has been removed
 - **grunt-spritesmith** has been removed
 - **interactjs** has been upgraded from 1.9.7 to **1.10.11**
 - **node-schedule** has been upgraded from 1.3.2 to **2.0.0**